@@ -17,61 +17,60 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.0.0",
-    "@angular/common": "^19.0.0",
-    "@angular/compiler": "^19.0.0",
-    "@angular/core": "^19.0.0",
-    "@angular/forms": "^19.2.3",
-    "@angular/localize": "^19.2.3",
-    "@angular/platform-browser": "^19.0.0",
-    "@angular/platform-browser-dynamic": "^19.2.3",
-    "@angular/router": "^19.2.3",
-    "@ngxs/router-plugin": "^19.0.0",
-    "@ngxs/storage-plugin": "^19.0.0",
-    "@ngxs/store": "^19.0.0",
-    "@primeng/themes": "^19.0.10",
+    "@angular/animations": "^20.1.0",
+    "@angular/common": "^20.1.0",
+    "@angular/compiler": "^20.1.0",
+    "@angular/core": "^20.1.0",
+    "@angular/forms": "^20.1.1",
+    "@angular/localize": "^20.1.1",
+    "@angular/platform-browser": "^20.0.0",
+    "@angular/platform-browser-dynamic": "^20.1.1",
+    "@angular/router": "^20.1.1",
+    "@ngxs/router-plugin": "^20.1.0",
+    "@ngxs/storage-plugin": "^20.1.0",
+    "@ngxs/store": "^20.1.0",
+    "@primeng/themes": "^20.0.0",
     "crypto-js": "^4.2.0",
     "fast-json-patch": "^3.1.1",
     "moment": "^2.30.1",
     "primeflex": "^4.0.0",
     "primeicons": "^7.0.0",
-    "primeng": "^19.0.10",
+    "primeng": "^20.0.0",
     "rxjs": "^7.8.2",
     "tslib": "^2.8.1",
-    "zone.js": "^0.15.0"
+    "zone.js": "^0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.2.4",
-    "@angular/cli": "^19.2.4",
-    "@angular/compiler-cli": "^19.0.0",
-    "@angular/language-service": "^19.2.3",
-    "@badeball/cypress-cucumber-preprocessor": "^22.0.1",
-    "@bahmutov/cypress-esbuild-preprocessor": "^2.2.4",
+    "@angular/build": "^20.1.1",
+    "@angular/cli": "^20.1.1",
+    "@angular/compiler-cli": "^20.1.0",
+    "@angular/language-service": "^20.1.1",
+    "@badeball/cypress-cucumber-preprocessor": "^22.2.0",
+    "@bahmutov/cypress-esbuild-preprocessor": "^2.2.5",
     "@cypress/schematic": "^3.0.0",
     "@sinonjs/fake-timers": "^14.0.0",
-    "@types/chai": "^5.2.1",
+    "@types/chai": "^5.2.2",
     "@types/mocha": "^10.0.10",
-    "@types/node": "^22.13.13",
+    "@types/node": "^22.16.4",
     "@types/sinon": "^17.0.4",
-    "@typescript-eslint/eslint-plugin": "^8.21.0",
-    "@typescript-eslint/parser": "^8.21.0",
-    "angular-eslint": "^19.3.0",
+    "@typescript-eslint/eslint-plugin": "^8.37.0",
+    "@typescript-eslint/parser": "^8.37.0",
+    "angular-eslint": "^20.1.1",
     "cassandra-driver": "^4.8.0",
-    "chai": "^5.2.0",
-    "cypress": "^14.2.0",
-    "eslint": "^9.23.0",
+    "chai": "^5.2.1",
+    "cypress": "^14.5.2",
+    "eslint": "^9.31.0",
     "karma": "^6.4.4",
     "karma-chrome-launcher": "^3.2.0",
     "karma-coverage": "^2.2.1",
     "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
-    "mocha": "^11.1.0",
-    "ng-mocks": "^14.13.4",
-    "sinon": "^19.0.4",
+    "mocha": "^11.7.1",
+    "ng-mocks": "^14.13.5",
+    "sinon": "^21.0.0",
     "sonarqube-scanner": "^4.3.0",
-    "ts-loader": "^9.5.2",
-    "typescript": "^5.8.2",
-    "typescript-eslint": "^8.28.0"
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.37.0"
   },
   "cypress-cucumber-preprocessor": {
     "nonGlobalStepDefinitions": true,
