@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, inject } from '@angular/core';
 import { FormsModule, ReactiveFormsModule, UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { Store } from '@ngxs/store';
 import { ButtonModule } from 'primeng/button';
@@ -29,12 +29,11 @@ import { notBlank } from '../../shared/validator/not-blank.validator';
 })
 export class AccountCreateComponent implements OnInit {
 
-  accountForm: UntypedFormGroup;
+  private readonly fb = inject(UntypedFormBuilder);
+  private readonly store = inject(Store);
+  private readonly loginValidator = inject(LoginValidator);
 
-  constructor(
-    private readonly fb: UntypedFormBuilder,
-    private readonly store: Store,
-    private readonly loginValidator: LoginValidator) { }
+  accountForm: UntypedFormGroup;
 
   ngOnInit() {
 
