@@ -1,5 +1,5 @@
 import { HttpStatusCode, HttpClient, HttpErrorResponse, HttpHeaders, HttpResponse } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
 import { Observable, of, throwError, from } from 'rxjs';
 import { catchError, map, mergeMap, takeWhile } from 'rxjs/operators';
 
@@ -8,7 +8,7 @@ import { Login } from './login';
 @Injectable({ providedIn: 'root' })
 export class LoginService {
 
-  constructor(private readonly http: HttpClient) { }
+  private readonly http = inject(HttpClient);
 
   checkUsername(username: string): Observable<boolean> {
 
