@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
 import { Action, State, StateContext } from '@ngxs/store';
 import { tap } from 'rxjs/operators';
 
@@ -12,8 +12,7 @@ import { AuthService } from '../auth/shared/auth.service';
 @Injectable()
 export class AppState {
 
-  constructor(
-    private readonly authService: AuthService) { }
+  private readonly authService = inject(AuthService);
 
   @Action(Authenticate)
   authenticate(ctx: StateContext<boolean>) {
