@@ -1,14 +1,12 @@
 import { Routes } from '@angular/router';
 
 import { AnonymousGuard, AuthenticatedGuard } from '../auth/shared/auth.guard';
-import { AccountCreateComponent } from './account-create/account-create.component';
-import { AccountEditComponent } from './account-edit/account-edit.component';
 import { AccountResolver } from './shared/account.resolver';
 
 export const ACCOUNT_ROUTES: Routes = [
   {
     path: '',
-    component: AccountEditComponent,
+    loadComponent: () => import('./account-edit/account-edit.component').then(m => m.AccountEditComponent),
     canActivate: [
       AuthenticatedGuard
     ],
@@ -18,7 +16,7 @@ export const ACCOUNT_ROUTES: Routes = [
   },
   {
     path: 'create',
-    component: AccountCreateComponent,
+    loadComponent: () => import('./account-create/account-create.component').then(m => m.AccountCreateComponent),
     canActivate: [
       AnonymousGuard
     ]
