@@ -1,12 +1,11 @@
 import { Routes } from '@angular/router';
 
 import { AnonymousGuard } from './shared/auth.guard';
-import { AuthLoginComponent } from './auth-login/auth-login.component';
 
 export const AUTH_ROUTES: Routes = [
   {
     path: '',
-    component: AuthLoginComponent,
+    loadComponent: () => import('./auth-login/auth-login.component').then(m => m.AuthLoginComponent),
     canActivate: [
       AnonymousGuard
     ]
